requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
pyarrow==14.0.1
websockets==12.0
aiohttp==3.9.1
httpx==0.25.2
jinja2==3.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-binance==1.0.19
numpy==1.25.2
scipy==1.11.4
ta==0.10.2
